fix(validation): change Component field to pointer in VisibleForEntityContext (platform-mesh/extension-manager-operator#240)
The component field is being added because it is part of the VisibleForEntityContext struct in the model.go file, and its omitempty tag only prevents it from being included in the JSON output if it is nil or empty. However, since component is a struct (not a pointer), it is always initialized with its zero value (an empty Component struct), which is not considered "empty" by the omitempty tag.

On-behalf-of: @SAP angel.kafazov@sap.com
Signed-off-by: Angel Kafazov <akafazov@cst-bg.net>